Update ancestor folders to the current time when a file is deleted.
Currently, when a file is deleted, the remaining folders have their
timestamp set to the maximum timestamp of any of their remaining
children. However, this means that, for example:

- At T=1, dir/a.txt is written
- At T=2, dir/b.txt is written
- At T=3, dir/a.txt is deleted

Currently, dir will have version '2' after these changes. A client that
has synced up to T=2 will not see dir's version change at T=3 and will
therefore not know a.txt has been deleted. The current versioning
masks any deletion of files that are not the most recently changed in
their tree.

This change sets all ancestor folders to the current time, which should
be later than all other changes, so that dir has version '3' after the
above operations.

This would stand a better chance of being correct if we used hash trees,
but we're not there yet.